+++ scripts/alignmentIntervals.py
